Moved state pattern into different files
