+++ proxy/README.md
@@ -1,0 +1,17 @@
+# The Proxy Pattern
+
+## Intent
+
+## Structure
+
+![](../data/proxy_pattern_uml.png) 
+
+## Terminology
+
+| Term                     | Meaning                                                                                                                                         |
+| ------------------------ | ----------------------------------------------------------------------------------------------------------------------------------------------- |
+
+## Explanation
+
+## Applicability 
+
